DOS_PROJECT/requirements.txt
asgiref==3.6.0
autopep8==2.0.2
backports.zoneinfo==0.2.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
defusedxml==0.7.1
Django==4.2
django-allauth==0.54.0
idna==3.4
oauthlib==3.2.2
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
requests==2.29.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
tomli==2.0.1
urllib3==1.26.15
